Rename de algunas clases
