+++ gui.py
@@ -9,10 +9,10 @@ from PIL import Image
 from PIL import ImageTk
 
 import constants.colors as color
-from components.BrightnessContrastDialog import BrightnessContrastDialog
-from components.AutoScrollbar import AutoScrollbar
-from utils.utils import openCVToPIL, PILtoOpenCV
-from VeinSegmentation import mask
+from Components.BrightnessContrastDialog import BrightnessContrastDialog
+from Components.AutoScrollbar import AutoScrollbar
+from Utils.Utils import openCVToPIL, PILtoOpenCV
+from VeinSegmentation import Mask
 
 drawing = False
 ftypes = [('Imagen', '.png .jpeg .jpg')]
@@ -243,7 +243,7 @@ class App(Frame):
 
     def enhance(self):
         if len(self.polygon_points) > 1:
-            self.enhanced = mask.apply_enhance_to_roi(cv2.cvtColor(self.opencv_image, cv2.COLOR_GRAY2RGB), self.mask)
+            self.enhanced = Mask.apply_enhance_to_roi(cv2.cvtColor(self.opencv_image, cv2.COLOR_GRAY2RGB), self.mask)
             pts = np.array(self.polygon_points).reshape((-1, 1, 2))
             image_with_polygon = cv2.polylines(self.enhanced, [pts.astype(np.int32)], isClosed=self.isClosed,
                                                color=color.red, thickness=self.thickness)
@@ -257,9 +257,9 @@ class App(Frame):
     def skeletonize(self):
         if len(self.polygon_points) > 1:
             if self.is_enhanced:
-                self.skeletonized = mask.apply_skeletonization_to_roi(self.enhanced, self.mask, is_enhanced=True)
+                self.skeletonized = Mask.apply_skeletonization_to_roi(self.enhanced, self.mask, is_enhanced=True)
             else:
-                self.skeletonized = mask.apply_skeletonization_to_roi(self.opencv_image, self.mask, is_enhanced=False)
+                self.skeletonized = Mask.apply_skeletonization_to_roi(self.opencv_image, self.mask, is_enhanced=False)
             pts = np.array(self.polygon_points).reshape((-1, 1, 2))
             image_with_polygon = cv2.polylines(self.skeletonized, [pts.astype(np.int32)], isClosed=self.isClosed,
                                                color=color.red, thickness=self.thickness)
@@ -274,9 +274,9 @@ class App(Frame):
     def subpixel(self):
         if len(self.polygon_points) > 1:
             if self.is_enhanced:
-                self.subpixel_image = mask.apply_subpixel_to_roi(self.enhanced, self.mask, is_enhanced=True)
+                self.subpixel_image = Mask.apply_subpixel_to_roi(self.enhanced, self.mask, is_enhanced=True)
             else:
-                self.subpixel_image = mask.apply_subpixel_to_roi(self.opencv_image, self.mask, is_enhanced=False)
+                self.subpixel_image = Mask.apply_subpixel_to_roi(self.opencv_image, self.mask, is_enhanced=False)
 
             pts = np.array(self.polygon_points).reshape((-1, 1, 2))
             image_with_polygon = cv2.polylines(self.subpixel_image, [pts.astype(np.int32)], isClosed=self.isClosed,
