+++ scripts/subpixel_script.py
@@ -4,7 +4,7 @@ import cv2
 import matplotlib.pyplot as plt
 from time import time
 
-from VeinSegmentation.enhance import anisodiff
+from VeinSegmentation.Enhance import anisodiff
 from subpixel_edges import init, subpixel_edges
 
 scaleX = 1
